pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "api-debug-env"
version = "1.0.0"
description = "OpenEnv environment for AI-powered HTTP API debugging"
requires-python = ">=3.10"
dependencies = [
    "openenv-core>=0.2.3",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-asyncio>=0.21"]

[tool.setuptools.packages.find]
where = ["."]
include = ["server*", "tasks*", "graders*", "models*", "client*"]

[tool.uv.scripts]
server = "uvicorn server.app:app --host 0.0.0.0 --port 8000"